requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
base58==2.0.1
base91==1.0.1
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-spinner==0.1.10
colorama==0.4.4
commonmark==0.9.1
comtypes==1.1.7
Django==3.1.1
docopt==0.6.2
Faker==4.14.0
flake8==3.8.4
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
html5lib==1.1
idna==2.10
importlib-metadata==2.0.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
loguru==0.5.3
lxml==4.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
newsapi==0.1.1
newsapi-python==0.2.6
num2words==0.5.10
numerizer==0.1.5
numpy==1.19.2
paramiko==2.7.2
passlib==1.7.2
pgadmin4==4.27
psutil==5.7.3
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pandas==1.1.2
pybase62==0.4.3
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.7.1
pylint==2.6.0
pyshorteners==1.0.1
python-dateutil==2.8.1
pyttsx3==2.90
pytube3==9.6.4
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rich==8.0.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.10
webencodings==0.5.1
Werkzeug==1.0.1
win32-setctime==1.0.3
word2number==1.1
wrapt==1.12.1
yaspin==1.1.0
zipp==3.3.1
